flask-backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
supabase==2.15.0
python-dotenv==1.1.0
gunicorn==23.0.0
